// TheRiceMill/ClientApp/src/app/shared/model/create-rate-cost.model.ts
export class CreateRateCost {
    lotId: number;
    lotYear: number;
    labourUnloadingAndLoading: number;
    freight: number;
    purchaseBrokery: number;
    total: number;
    ratePer40WithoutProcessing: number;
    processingExpense: number;
    bardanaAndMisc: number;
    grandTotal: number;
    ratePer40LessByProduct: number;
    saleBrockery: number;
}

export class RateCost {
    id: number;
    lotId: number;
    lotYear: number;
    labourUnloadingAndLoading: number;
    freight: number;
    purchaseBrokery: number;
    total: number;
    ratePer40WithoutProcessing: number;
    processingExpense: number;
    bardanaAndMisc: number;
    grandTotal: number;
    ratePer40LessByProduct: number;
    saleBrockery: number;
}